requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
async-timeout==5.0.1
attrs==25.1.0
discord.py==2.4.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.0
python-dotenv==1.0.1
rarfile==4.2
typing-extensions==4.12.2
yarl==1.15.2
